users-backend/requirements.txt
aiomysql==0.1.1
anyio==3.6.2
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
click==8.1.3
cryptography==38.0.3
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.87.0
fastapi-users==10.2.1
fastapi-users-db-sqlalchemy==4.0.3
greenlet==2.0.1
h11==0.12.0
httpcore==0.15.0
httptools==0.5.0
httpx==0.23.0
httpx-oauth==0.10.2
idna==3.4
makefun==1.15.0
passlib==1.7.4
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
PyMySQL==1.0.2
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.21.0
typing-extensions==4.4.0
uvicorn==0.19.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
